Refactoring of previous work - Changes in API so users can select CodeMirror theme - bug fixes
