+++ js/src/main/scala/kazari/model/EvaluatorConfig.scala
@@ -3,4 +3,4 @@ package kazari.model
 import scala.scalajs.js.annotation.JSExport
 
 @JSExport
-case class EvaluatorConfig(url: String, authToken: String)
+case class PluginConfig(url: String, authToken: String, theme: String)
